Moving incident combobox to fixed incident filter combobox
