+++ src/dispatch/static/dispatch/src/task/TableFilterDialog.vue
@@ -41,7 +41,7 @@
 import { sum } from "lodash"
 import { mapFields } from "vuex-map-fields"
 
-import IncidentCombobox from "@/incident/IncidentCombobox.vue"
+import IncidentCombobox from "@/incident/IncidentFilterCombobox.vue"
 import IncidentPriorityCombobox from "@/incident/priority/IncidentPriorityCombobox.vue"
 import IncidentTypeCombobox from "@/incident/type/IncidentTypeCombobox.vue"
 import ProjectCombobox from "@/project/ProjectCombobox.vue"
